requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
deepl>=1.18.0
deep-translator>=1.11.4
